refactoring and reorganization for module and test management
